introducing a factory method to create instances of GetGpcStructuredTaskDefinition class
